+++ src/Random.php
@@ -24,31 +24,15 @@
 
 namespace fkooman\OAuth\Client;
 
-use ParagonIE\ConstantTime\Hex;
-
 class Random implements RandomInterface
 {
     /**
-     * Get the random data as raw binary string.
-     *
-     * @param int $length
+     * Return 256 bits "raw" random string.
      *
      * @return string
      */
-    public function getRaw($length)
+    public function raw()
     {
-        return \random_bytes($length);
-    }
-
-    /**
-     * Get the random data a hex encoded string.
-     *
-     * @param int $length
-     *
-     * @return string
-     */
-    public function getHex($length)
-    {
-        return Hex::encode($this->getRaw($length));
+        return \random_bytes(32);
     }
 }